requirements.txt
zhipuai>=2.0.0
python-dotenv>=1.0.0
sniffio>=1.3.0
httpx>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
